importer: assign transactions during ItemsWithoutAccountError exception in case only one account exists
In case only one account exists and transaction items are not flagged with an account iban, assign automatically all transactions to this one account.

The user will be notified with a flash message, that this assignment were done by MoneyWatch itself. In case these transactions are not related to the existing account, the user should create the new account first and then import these transactions again. In that case, the user will be prompted for which account these transactions should be imported.