Update README.md
Adapt plugin documentation to reflect the latest changes regarding account detection. The `account` keyword is no longer mandatory. In case no account IBAN can be given by an import plugin. MoneyWatch tries to autodetect the account if any transaction is already existing in the database. If no transaction can be found in the database during import, the user must decide for which account the transactions should be imported.