importer: show check/save button during import also on the bottom
During the edit of transactions to be imported, the check and save button are only shown on the top of the list. When importing a more than just a few seconds you have to scroll all the way up to use the buttons.

Now the buttons are also shown on the bottom if the numer of displayed transactions is greather than 3 transactions.