transactions: show always transactions
Currently only the transactions for the current month were shown. In case within this timeframe are no transactions existing, the list and also the search form will not be displayed and therefore historical transactions cannot be examined.

Now the date of the latest transaction will be taken and from there on the last 2 months will be displayed