analysis.py: substract the first item when displaying absolute account balance
This is neccessary due to the fact that otherwise, the account balance of the first (oldest) item in the dataset would reflect the overall sum, although only the account balance at the end of the period is displayed.

Without this, the difference between the first item account balance and the last item account balance would not match.