functions: fix regression bug when using 28th date with add_months/substract_months
When using add_months() or substract_months and using  29.01.2020  / 29.03.2020, the resulting date will return 30./31. as day result. This is due to the 29th is the last date of February. The current implementation only uses the date of the last calculating step to check, if the last day of month must be used or not. Now the original date will be used for this instead of the last calculating step.