importer: corrected trend calculation in some race conditions
The immediate trend calculation for transactions to be imported was based on the last transaction stored in the database.
In case the importing file contains several intervals, all transaction trends were calculated based on the latest transaction stored in the database but thats wrong in such cases.
Each transaction should be use the latest transaction before no matter if it is located in the database or is part of the running import process.
The trend calculation now considers also transactions that are part of the import when calculation the trend valuta.