conftest.py: always rollback a db session in postamble
When performing a pytest unit test using the db fixture it is possible in some unforseenable cases, that a db session has pending items that can't be commited (e.g. IntegrityError due to constraints). Therefore the db session must be rollback to ensure no pending items are leftover and avoid exceptions during commit. As the db is just temporary for unit tests, no need to commit anyway,