ruleset: make short description for new rules optional
Currently when creating a rule, the rule name must be given and in addition the short description which will be used for matched transactions as default value during import.

In most cases, the rule name is equal to the short description, therefore the short description will be made optional. If not explicitly given, the rule name will be used as descpription for newly found transactions during import.