ruleset: make rules disabled instead of deleting them
Currently when a rule is deleted, it is permanently deleted and all references to this rule (transactions) are lost.

To be able to still see the rule in the analysis as long as related transactions exist, the deleted rule will not really deleted but gets into a "disabled" state, where the rule still is present in the database but only for analysis reasons. Those disabled rules will not be shown in the ruleset overview and cannot be edited anymore. Disabled rules will not be used during import. This new behavior can be enabled/disabled via a new config key "KEEP_DELETED_OBJECTS_FOR_ANALYSIS" it defaults to True